+++ src/main/java/core/render360/Render360Event.java
@@ -1,0 +1,5 @@
+package core.render360;
+
+public class Render360Event {
+
+}
