+++ src/main/java/core/render360/coretransform/classtransformers/ClassTransformer.java
@@ -1,12 +1,12 @@
-package mod.render360.coretransform.classtransformers;
+package core.render360.coretransform.classtransformers;
 
 import org.apache.commons.lang3.ArrayUtils;
 import org.objectweb.asm.tree.ClassNode;
 import org.objectweb.asm.tree.MethodNode;
 
-import mod.render360.coretransform.classtransformers.name.ClassName;
-import mod.render360.coretransform.classtransformers.name.MethodName;
-import mod.render360.coretransform.classtransformers.particle.ParticleTransformer;
+import core.render360.coretransform.classtransformers.name.ClassName;
+import core.render360.coretransform.classtransformers.name.MethodName;
+import core.render360.coretransform.classtransformers.particle.ParticleTransformer;
 
 /**
  * Holds all of the class transformers.
