+++ src/main/java/mod/render360/gui/Render360Settings.java
@@ -6,7 +6,7 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
 
-import mod.render360.coretransform.RenderUtil;
+import core.render360.coretransform.RenderUtil;
 import mod.render360.render.RenderMethod;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.GuiScreen;
